user controller and book controller safety

+++ components/books/controller.js
@@ -1,9 +1,10 @@
 const modelBook = require("./model")
 
 
+
 const getBooks = async () => {
     try {
-        const response = await modelBook.find().populate("user","name")
+        const response = await modelBook.find().populate("user", "name")
         return response
     } catch (error) {
         throw new Error(error)
@@ -12,10 +13,13 @@ const getBooks = async () => {
 
 const addBook = async (name, gender) => {
     try {
+        if (!name || !gender) {
+            throw new Error("informacion incorrecta")
+        }
         const newBook = new modelBook({
             name,
             gender,
-            status:false
+            status: false
         })
         const response = await newBook.save()
         return response
@@ -26,7 +30,7 @@ const addBook = async (name, gender) => {
 
 const getBook = async (id) => {
     try {
-        const bookSelected = await modelBook.findOne({ _id: id }).populate("user")
+        const bookSelected = await modelBook.findById(id).populate("user")
         return bookSelected
     } catch (error) {
         throw new Error(error)
@@ -36,6 +40,9 @@ const getBook = async (id) => {
 
 const updateBook = async (id, name, gender) => {
     try {
+        if (!name || !gender || id) {
+            throw new Error("informacion incorrecta")
+        }
         const bookSelected = await modelBook.findOne({ _id: id })
         bookSelected.name = name
         bookSelected.gender = gender
@@ -48,7 +55,15 @@ const updateBook = async (id, name, gender) => {
 
 const deleteBook = async (id) => {
     try {
+
+        const bookSelected = await modelBook.findById(id)
+
+        if (bookSelected.status) {
+            throw new Error("existe usuario con este libro")
+        }
+
         await modelBook.deleteOne({ _id: id })
+
         return "registro borrado con exito"
     } catch (error) {
         throw new Error(error)
@@ -57,7 +72,7 @@ const deleteBook = async (id) => {
 
 const deleteAllBooks = async () => {
     try {
-        await modelBook.deleteMany()
+        await modelBook.deleteMany({status:false})
         return "Registros borrados con exito"
     } catch (error) {
         throw new Error(error)
@@ -71,7 +86,7 @@ module.exports = {
     getBook: getBook,
     updateBook: updateBook,
     deleteBook: deleteBook,
-    deleteAllBooks:deleteAllBooks
+    deleteAllBooks: deleteAllBooks
 }
